requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.5
whitenoise==6.11.0
